# packages/protocol/src/generated/openapi.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComaMessenger API",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/push/subscriptions": {
      "put": {
        "operationId": "putPushSubscription",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered browser subscription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscription"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/push/subscriptions/{subscription_id}": {
      "delete": {
        "operationId": "deletePushSubscription",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Subscription removed."
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/preferences": {
      "get": {
        "operationId": "getPreferences",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Client preferences.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updatePreferences",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated preferences.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Error": {
        "description": "Stable API error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "already_bootstrapped",
          "chat_conflict",
          "chat_not_found",
          "forbidden",
          "idempotency_conflict",
          "internal_error",
          "invalid_credentials",
          "invalid_refresh_token",
          "invalid_request",
          "invitation_invalid",
          "message_not_found",
          "origin_not_allowed",
          "payload_too_large",
          "rate_limited",
          "service_not_ready",
          "session_not_found",
          "unauthorized",
          "unsupported_format",
          "validation_failed",
          "version_conflict"
        ]
      },
...
      "PushSubscriptionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "endpoint",
          "keys"
        ],
        "properties": {
          "endpoint": {
            "type": "string",
            "format": "uri"
          },
          "keys": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "p256dh",
              "auth"
            ],
            "properties": {
              "p256dh": {
                "type": "string"
              },
              "auth": {
                "type": "string"
              }
            }
          }
        }
      },
      "PushSubscription": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "endpoint",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "endpoint": {
            "type": "string",
            "format": "uri"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserPreferences": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "theme",
          "locale",
          "push_enabled",
          "push_preview"
        ],
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "system",
              "light",
              "dark"
            ]
          },
          "locale": {
            "type": "string",
            "enum": [
              "ru",
              "en"
            ]
          },
          "push_enabled": {
            "type": "boolean"
          },
          "push_preview": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}